requirements.txt
autopep8==1.5.2
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.1
chardet==3.0.4
click==7.1.1
cryptography==3.0
flake8==3.8.3
Flask==1.1.2
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Injector==0.12.3
Flask-Mail==0.9.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
injector==0.18.3
itsdangerous==1.1.0
Jinja2==2.11.2
jsonrpclib==0.1.7
jsonrpclib-pelix==0.4.1
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
mccabe==0.6.1
ndg-httpsclient==0.5.1
passlib==1.7.2
pusher==3.0.0
pyasn1==0.4.8
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyotp==2.4.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
typing==3.7.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.3
numpy==1.19.1
pandas==1.0.5
